changed github users to the about me page
