+++ frontend/src/app/app.component.html
@@ -2,3 +2,4 @@
 <main class="content">
   <router-outlet></router-outlet>
 </main>
+<app-footer/>
